pyproject.toml
[project]
name = "ai-banking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0",
    "langchain>=1.0.3",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "plotly>=6.3.1",
    "pyarrow==16.1.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
